Created family class
I'm trying to get what i have to do to make it work. Because if every guest ,from the same family, enters a room at the same time and they also leave together, then it would be better to a family be  a Thread instead of each guest be Thread.